Cleanup
Put sub-resources classes and schemas in dedicated package.
This is most consistent with the way schemas and resources are stored in the other plugins.
